suite

+++ ft_strncmp.c
@@ -14,13 +14,13 @@
 /* ************************************************************************** */
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int i;
-	int nn;
-
+	size_t i;
+	size_t nn;
 
 	if (n == 0)
 		return (0);
-	nn = (int)n - 1;	
+
+	nn = n - 1;	
 	
 	i = 0;
 	while (1)
@@ -34,5 +34,4 @@ int	ft_strncmp(const char *s1, const char *s2, size_t n)
 	}
 
 	return ( (unsigned char)s1[i] - (unsigned char)s2[i]);
-
 }
